+++ history.md
@@ -2,6 +2,8 @@ History
 -------
 In reversed order.
 
+* Add yaml properties for h2 and mysql.
+* Use mysql_db_users.sql script to create DB and user.
 * UserPrincipal implements UserDetails.
 * HttpResponse - A uniform way of sending a response back to the client.
 * User entity.